Finish match of grlib, fix fabsf declaration, fixup Fighter's x1860 naming based on usages, improved WIPs in ftCo_DamageFall and ftwaitanim (#1441)
* Get ftwaitanim closer

* Fix fabsf declaration

* Get grLib_801C9EE8 match closer

* Fixup usages of x1860 to reflect as the dmg element

* WIP

* WIP 2

* Match grLib_801C9EE8 thanks to foxcam

* Mark grlib as matching

* Correction to fabsf__Ff call

* Tweak cast

* Comment out cast for now due to error

+++ src/MSL/math.h
@@ -85,7 +85,7 @@ static inline f32 fabs_inline(f32 x)
 }
 
 double frexp(double x, int* exponent);
-double fabsf__Ff(double);
+float fabsf__Ff(float);
 float tanf(float x);
 float cos__Ff(float x);
 float sin__Ff(float x);
